requirements.txt
streamlit
pandas
numpy
tensorflow
requests
scikit-learn
plotly
python-dotenv